The basic moodle plugins for publishing video
filtersettings.php needs some work to make the filter easier to install
in a non-illinois.edu moodle server. Similarly, the admin mforms in the
repo lib.php should be made more robust to more intelligently acquire a
destinationID (currently an admin has to paste it in). Filter is updated
to make an API call to determine correct video dimensions. Repo
thumbnails aren't formatted correctly--I had previously modified
filepicker.js to fix that behavior on my dev sandbox, but that's not the
right way to go.